refactor: move tests to the assembler
