+++ miden/tests/integration/main.rs
@@ -7,7 +7,6 @@ mod cli;
 mod exec_iters;
 mod flow_control;
 mod operations;
-mod program;
 
 // TESTS
 // ================================================================================================
